Fix bug with var access within closure
The gofunc() closure within AllInspectedContainers() is referencing
shortDescContainer.ID from the enclosing scope. Because
shortDescContainer.ID is inside of a loop, its value is unpredicable
while the goroutines are running. The usual effect is that all
goroutines get the same value for the ID. The solution is to set
the ID in the larger container before starting the goroutine.

Signed-off-by: Bill Green <bgreen@newrelic.com>