Use the actual name for the container as hash keys for clarity
Use this:

m.Data[container.Names[0]] = data

Instead of:

m.Data[container.Image+"-"+container.ID]

For the json structures returned by:

/r/docker/containers/cpu
/r/docker/containers/memory
/r/docker/containers/net/dev
/r/docker/containers"

The container name is used in other services to identify the container,
use it in place of image + ID. This makes it easier to line up the
status from resourced to output from mesos, sidecar, etc.

Docker requires that container names be unique, so this should do for a
hash key. Are there cases where the name could be null?

Signed-off-by: Bill Green <bgreen@newrelic.com>